Fixing bugs in parallel:
- wrong communications when one proc has no nodes to send
- globalToLocalEquationNumber in applyBoundary
- adding possibility to specify the seed in the command line

+++ src/io/parser/cppargparse/cppargparse.cc
@@ -138,7 +138,7 @@ ArgumentParser::_addArgument(const std::string & name, const std::string & help,
     break;
   }
   case _integer: {
-    arg = new ArgumentStorage<int>();
+    arg = new ArgumentStorage<long int>();
     break;
   }
   case _boolean: {
@@ -406,7 +406,7 @@ bool ArgumentParser::checkType(ArgumentType type,
     break;
   }
   case _integer: {
-    int i;
+    long int i;
     sstr >> i;
     break;
   }
